Create Model interface to abstract mat specifics
I then created Phong implementations of this interface to serve the
existing "Standard" materials (which have since been renamed to Phong to
clarify things). This is so that I did not have to rewrite/scrap code to
create Pbr models, and they will still have the efficincies that I
worked hard to implement for model loading.